src/app/requirements.txt
flask
pandas
mlflow
python-dotenv
scikit-learn
tabpfn